CLN: initial commit with incoperation of task_toggle_group_accelerator.py into wfmanager.py
