+++ force_wfmanager/ui/variable_names_registry.py
