+++ force_wfmanager/wfmanager.py
@@ -4,6 +4,9 @@ from envisage.ui.tasks.api import TasksApplication, TaskWindow
 
 from pyface.api import (CANCEL, ConfirmationDialog, NO, YES)
 from pyface.tasks.api import TaskWindowLayout
+from pyface.action.api import ActionItem
+from pyface.tasks.action.api import TaskToggleGroup
+from pyface.tasks.action.task_toggle_group import TaskToggleAction
 
 from traits.api import Either, Int, Tuple
 
@@ -124,3 +127,15 @@ class TaskWindowClosePrompt(TaskWindow):
         # Cancel
         elif result is CANCEL:
             return False
+
+
+class TaskToggleGroupAccelerator(TaskToggleGroup):
+    """The standard TaskToggleGroup from pyface, but with a keyboard shortcut
+    for changing tasks (Ctrl + index of task)."""
+    def _get_items(self):
+        items = []
+        for i, task in enumerate(self.window.tasks):
+            action = TaskToggleAction(task=task,
+                                      accelerator='Ctrl+{}'.format(i+1))
+            items.append(ActionItem(action=action))
+        return items
